requirements.txt
torch==2.2.0+cu121
torchvision==0.17.0+cu121
numpy==1.24.3
pandas==2.2.3
pyyaml==6.0.2
open-clip-torch==2.28.0
Pillow==10.4.0
